requirements.txt
Flask==0.10.1
Flask-API==0.6.2
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
nose==1.3.4
wsgiref==0.1.2
